MSVC touchups: Added snmptrapd project .DSP file; All .DS[PW] files CR-LF line terminated; Fix warnings in asn1.c and snmp_api.c; Remove unused var in snmptranslate.c.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk/net-snmp@3382 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/asn1.c
@@ -149,7 +149,7 @@ asn_check_packet (u_char *pkt, size_t len) {
 
   if (*(pkt+1) & 0x80) {
     /* long length */
-    if (len < (*(pkt+1) & ~0x80)+2)
+    if ((int)len < (int)(*(pkt+1) & ~0x80)+2)
       return 0;    /* still to short, incomplete length */
     asn_parse_length(pkt+1, &asn_length);
     return (asn_length + 2 + (*(pkt+1) & ~0x80));
